requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
numpy==1.16.2
Pillow==7.0.0
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4
pyxdg==0.26
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
urllib3==1.24.2
webcolors==1.10
